Fix the "nil bug" in `clause` implementation
`clause` invoked the action, when the pattern matched. The problem with
this is that when it returns `nil`, it's impossible to know if the
pattern did not match, or whether the action returned `nil`. We assumed
the former and fell through to the next pattern. Now we cannot even
prohibit actions from returning `nil` because it's a totally legit
thing to do.

Hence we now return a different token - ::clause-not-applied - to
indicate when the clause could not be applied.