Add guard pattern syntax
These are not full-fledged pattern guards that you see in ML and the
likes. They can refer to arbitrary bindings, and will require a
significant change to Akar's core model. I instead decided to something
simpler, and only allow point-free functions to be passed to be applied
to the value being matched. This should be good enough. (Teehee!) A
precedent can be found in core.match, who took the same approach.