Add syntax support for or-patterns
The underlying !or combinator is capable of emitting values. However
it is not possible to expose this feature via syntactic :or. It is
restricted to non-emitting patterns. Note that our syntax rule is not
capable of determining whether or not there will be emissions from a
pattern. Hence it is up to the user to only use non-emitting patterns
from within :or.