Add syntactic support for non-literal constant patterns
There was no dedicated syntax in Akar for matching against non-literal
constants. Most pattern implmenetations have it, because it's so often
needed.

Because of Akar's first-classiness, one could still do this:

(let [x 11]
  (match 11
         [(!constant x)] :we-will-get-here))

With the new syntax, it's slightly more concise:

(let [x 11]
  (match 11
         (:constant x) :we-will-get-here))

This syntax is analogous to backticks in patterns in Scala.

val x = 11
11 match {
  case `x` => 'weWillGetHere
}