Move over useful code from 'gophan'
Original commit message from 2016.07.10

* Add utilities to write macros using clojure.spec

I am unsure this is the right way to do it, but I could not figure out
or find anything better.

Apparently, you can write a spec for a macro, and macroexpansion
automatically performs the checks for you, but "validation is parsing"
bit is not obvious here. Once the macro expander has invoked `conform`,
the results of that operation i.e. the parsed data should be available
for further consumption, which it does not seem to be. Weird.

---

I had a Twitter conversation with Alex Miller at the time wherein I
asked him if there was a better way available. It seemed, at the time,
there wasn't. Not sure if there's anything better available today.