Stop binding a variable in `:type` pattern
I was binding a variable by default because this is how Scala does it:
Its typecasing support has binding enabled by default. Like so:
`ex: Exception`. You must ignore it explicitly if you don't need it.
Like so: `_: Exception`. In Akar, we have at/and patterns, something
Scala lacks. So we don't really need to muddle responsibilities like
this.

This is a breaking change.